@@ -12,13 +12,13 @@
     <artifactId>rookie-springboot-mybatis</artifactId>
 
     <name>rookie-springboot-mybatis</name>
-    <!-- FIXME change it to the project's website -->
+    <!--  change it to the project's website -->
     <url>http://www.example.com</url>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
     </properties>
 
     <dependencies>
