@@ -2,7 +2,7 @@ package com.rookie.bigdata.framework.stereotype.component;
 
 /**
  * @Class A
- * @Description TODO
+ * @Description
  * @Author rookie
  * @Date 2024/6/27 15:10
  * @Version 1.0
