@@ -2,7 +2,7 @@ package com.rookie.bigdata.retry.callback;
 
 /**
  * @Class RetryPolicy
- * @Description TODO
+ * @Description
  * @Author rookie
  * @Date 2024/1/2 14:13
  * @Version 1.0
