@@ -6,7 +6,7 @@ import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 
 /**
  * @Class AsyncRetryExecutor
- * @Description TODO
+ * @Description
  * @Author rookie
  * @Date 2024/1/2 14:19
  * @Version 1.0
