@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 /**
  * @Classname LockController
- * @Description TODO
+ * @Description
  * @Author rookie
  * @Date 2023/1/16 11:31
  * @Version 1.0
